[LICENSE] Moved lessons in a separate folder with a different License then the rest of the code
