+++ frontend-ssr/curriculum/intro/About.tsx
