+++ frontend-ssr/pages/intro/about.tsx
@@ -1,0 +1,1 @@
+export { default } from '~/curriculum/intro/About';
